New widget added, updateError method added
This is a summary of the latest commits that have been concatenated. An multitude of new features to the RapidReact-2022/RapidReact-2022/src/main/java/frc/robot/util/ShufflePID.java. An new widget have been added, used to display the error of the Proportional-Integral-Derivative (PID) field, bind. The widget has been configured to be displayed as an graph, as that would be the upmost efficient method for displaying change in the error. I have made a public method for the class, named updateError, which will return void. The function will promptly call the getError method of the Proportional-Integral-Derivative (PID) field, often referred to as bind. This method will henceforth return a double, which represents the prevError field of the Proportional-Integral-Derivative (PID) instance. Since the field is private, this method is required and recommended. We use the double returned by getError to update the graph using a setDouble method, which is built into the NetworkTableEntry object. Lastly, we call updateError() in the update() method. This would update the graph every time the method is called. This description has been translated into multiple languages.